requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
diffusers==0.6.0
filelock==3.8.0
huggingface-hub==0.10.1
idna==3.4
importlib-metadata==5.0.0
mpmath==1.2.1
networkx==3.0b1
numpy==1.23.4
packaging==21.3
Pillow==9.2.0
pyparsing==3.0.9
PyYAML==6.0
regex==2022.9.13
requests==2.28.1
sympy==1.11.1
tokenizers==0.13.1
torch==1.14.0.dev20221020
torchaudio==0.14.0.dev20221020
torchvision==0.15.0.dev20221020
tqdm==4.64.1
transformers==4.23.1
typing_extensions==4.4.0
urllib3==1.26.12
zipp==3.9.0
